tfrecords/requirements.txt
Cython
javaproperties
argparse
contextlib2
pillow
lxml
jupyter
matplotlib
numpy
